+++ lib/kubecontrol.rb
@@ -1,10 +1,6 @@
 require 'kubecontrol/version'
 require 'kubecontrol/client'
-require 'kubecontrol/pod'
-require 'kubecontrol/service'
-require 'kubecontrol/stateful_set'
-require 'kubecontrol/deployment'
-require 'kubecontrol/secret'
+require 'kubecontrol/resources'
 
 module Kubecontrol
   class Error < StandardError; end