+++ lib/kubecontrol/client.rb
@@ -1,9 +1,5 @@
 require 'open3'
-require_relative 'pod'
-require_relative 'deployment'
-require_relative 'stateful_set'
-require_relative 'service'
-require_relative 'secret'
+require_relative 'resources'
 
 module Kubecontrol
   class Client
@@ -26,23 +22,23 @@ module Kubecontrol
     end
 
     def pods
-      get_resource(Pod, 5)
+      get_resource(Resources::Pod, 5)
     end
 
     def deployments
-      get_resource(Deployment, 5)
+      get_resource(Resources::Deployment, 5)
     end
 
     def stateful_sets
-      get_resource(StatefulSet, 3)
+      get_resource(Resources::StatefulSet, 3)
     end
 
     def services
-      get_resource(Service, 6)
+      get_resource(Resources::Service, 6)
     end
 
     def secrets
-      get_resource(Secret, 4)
+      get_resource(Resources::Secret, 4)
     end
 
     def find_secret_by_name(name_regex)