+++ spec/kubecontrol/resources/pod_spec.rb
@@ -1,7 +1,7 @@
-require_relative '../spec_helper'
-require_relative '../../lib/kubecontrol/pod'
+require_relative '../../spec_helper'
+require_relative '../../../lib/kubecontrol/resources/pod'
 
-RSpec.describe Kubecontrol::Pod do
+RSpec.describe Kubecontrol::Resources::Pod do
   let(:pod_name) { 'foo_pod' }
   let(:pod_ready) { '1/1' }
   let(:pod_status) { 'Running' }
@@ -11,7 +11,7 @@ RSpec.describe Kubecontrol::Pod do
   let(:client) { Kubecontrol::Client.new }
 
   describe '#initialize' do
-    subject { Kubecontrol::Pod.new(pod_name, pod_ready, pod_status, pod_restarts, pod_age, namespace, client) }
+    subject { Kubecontrol::Resources::Pod.new(pod_name, pod_ready, pod_status, pod_restarts, pod_age, namespace, client) }
 
     it 'sets the pod name field' do
       expect(subject.name).to eq pod_name
@@ -43,7 +43,7 @@ RSpec.describe Kubecontrol::Pod do
   end
 
   describe '#running?' do
-    subject { Kubecontrol::Pod.new(pod_name, pod_ready, pod_status, pod_restarts, pod_age, namespace, client).running? }
+    subject { Kubecontrol::Resources::Pod.new(pod_name, pod_ready, pod_status, pod_restarts, pod_age, namespace, client).running? }
 
     context 'is running' do
       it { is_expected.to eq true }
@@ -57,7 +57,7 @@ RSpec.describe Kubecontrol::Pod do
   end
 
   describe '#stopped?' do
-    subject { Kubecontrol::Pod.new(pod_name, pod_ready, pod_status, pod_restarts, pod_age, namespace, client).stopped? }
+    subject { Kubecontrol::Resources::Pod.new(pod_name, pod_ready, pod_status, pod_restarts, pod_age, namespace, client).stopped? }
 
     context 'is running' do
       it { is_expected.to eq false }
@@ -77,7 +77,7 @@ RSpec.describe Kubecontrol::Pod do
     let(:std_err) { '' }
     let(:status_code) { 0 }
     let(:kubectl_command_response) { [std_out, std_err, status_code] }
-    let(:pod) { Kubecontrol::Pod.new(pod_name, pod_ready, pod_status, pod_restarts, pod_age, namespace, client) }
+    let(:pod) { Kubecontrol::Resources::Pod.new(pod_name, pod_ready, pod_status, pod_restarts, pod_age, namespace, client) }
 
     subject { pod.exec(command) }